+++ src/modules/api/spoonacular/spoonacular.api.service.ts
@@ -2,11 +2,16 @@ import { Injectable } from '@nestjs/common';
 import { DetailedMeal, RatedMeal } from '../../meal/meal.types';
 import { AbstractApiService } from '../../../services/abstract.api.service';
 import { ApiName } from '../../redis/redis.types';
-import { SpoonacularIngredient, SpoonacularRecipe, SpoonacularRecipeDetails } from './spoonacular.api.types';
+import {
+    SpoonacularIngredient,
+    SpoonacularMealInstruction,
+    SpoonacularRecipe,
+    SpoonacularRecipeDetails
+} from './spoonacular.api.types';
 import { IngredientType, MealIngredient } from '../../ingredient/ingredient.types';
 
 @Injectable()
-export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe, SpoonacularIngredient, SpoonacularRecipeDetails> {
+export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe, SpoonacularIngredient, SpoonacularRecipeDetails, SpoonacularMealInstruction> {
 
     getApiUrl(): string {
         return 'https://api.spoonacular.com';
@@ -41,7 +46,7 @@ export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe,
         });
     }
 
-    proceedDataToMealDetails(data: SpoonacularRecipeDetails): DetailedMeal {
+    proceedDataToMealDetails(data: SpoonacularRecipeDetails, instructionData: SpoonacularMealInstruction): DetailedMeal {
         const {
             id, image, title, extendedIngredients, summary,
             vegetarian, vegan, glutenFree, dairyFree, veryHealthy,
@@ -62,7 +67,8 @@ export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe,
                 glutenFree,
                 dairyFree,
                 veryHealthy
-            }
+            },
+            instruction: instructionData
         };
     }
 }